renamed arm_sim_control to arm_viz
